Refactor move DeleteSourceDialog tests
